requirements.txt
biopython==1.83
numpy==1.26.4
pyaml==24.4.0
PyYAML==6.0.1
